setup: Create folder structure for C++ client
Create folder structure for C++ client-side application for Windows platform. This commit includes README.md file in `docs/` folder. Also `.gitkeep` files are added in directories.

Initially only a single CPP file inside of the `src/` directory is provided with `main` function which is the entrypoint of the application. It still needs modification but this commit ensures a proper project folder structure to be there.

**Labels:** setup, documentation

Fixes #6